tools.experiments: unify scripts into 1
